Bugfix for generating operations
When generating code for PortType.Operations, we allocate an array of
size equals to the number of operations declared in the WSDL. However,
we skip operations that are not SOAP, like the RPC ones.

Therefore we have to trim this array when passing it to execute the
template to generate Go code.

Closes #16.