Add support for new types, and bug fixes
This commit adds support for token, anyURI, QName and unions as
described in the README.

The token and anyURI types are implemented as strings. The QName is
currently implemented as a string too but we might have to change it
whenever it causes trouble.

The union type is currently implemented as an empty interface with a
comment about the types accepted in the union. Later on we could have a
validator for that.

The "any" element of sequences in complex types is also supported now,
as a slice of empty interfaces, and the nonNegativeInteger type is
implemented as uint.

Restricted elements of complex types are still not implemented, the
likes of pattern matching (in a validation function) are lacking.

Functions in interfaces and methods in structs are now sorted by name.
This required a change in the memcache.golden file of unit tests.

One more important thing is that elements without a type are
automatically set to the string type. I dont know if this is right but
it works for now. yolo

Closes #20