Prevent element modification when dealing with complexType (#42)
* Prevent modification of originaly parsed element name for complexType

Name was modified through the reference. As a result tag had a wrong name
when the complexType was used multiple times.
With this update field generator operates on an element copy.

* Don't modify element type when generating element field

Element data should not be modified by a 'get' funtion. This may have a
side effect when the element is used multiple times.
To be safe, operate with the copy.

* Space stripped by go ftm