operation paramters: resolve element names
When an operation has a message that refers to a part whose element name
differs from it's element type, you need to try and resolve that name rather
than taking the type of the part directly.

This is the same name resolution as done when generating the operation
parameters elsewhere.